Add EIP-7951 Signer Support
Fixes #1054

This PR adds support for EIP-7951 (and thus RIP-7212) signers to the
Safe contracts, allowing P-256 signers to also sign Safe transactions.
In particular, this is interesting as it allows things like secure
enclave signatures (as the Nist P-256 curve is supported there while
the secp256k1 curve it not).

The new signer works by adding a new signature type (arbitrarily
chosen to be 2), which checks that the `p256verify` precompile returns
that a signature is valid, and checks `address(keccak256(qx, qy))` is
a Safe owner (same address derivation scheme as EOAs, but for the
secp256r1 curve).