Refactor Code for Formal Verification Support
We refactored the `P256VERIFY` precompile call into its own contract
and separate function in order to enable CVL summaries and allow the
FV to work. Otherwise, the FV will `havoc` on the call to the `0x100`
precompile (as the solver has no idea what can happen, and cannot
resolve the callee).

This makes only a very small (~30bytes) difference to codesize, and
marginal difference to gas costs, while improving code clarity; so it
is a worthwile change.